Enhancement #93 : Auto word complete searchbar extended
A user will get auto-suggestions depending on the characters typed. This feature has been extended. As soon as a user selects their chosen company and submits, details of that company will be displayed.
