+++ Project Code/Golang/NSU_alumni_portal/.idea/workspace.xml
@@ -11,8 +11,8 @@
   <component name="FileTemplateManagerImpl">
     <option name="RECENT_TEMPLATES">
       <list>
-        <option value="Go Application" />
         <option value="HTML File" />
+        <option value="Go Application" />
       </list>
     </option>
   </component>
@@ -73,7 +73,7 @@
       <breakpoints>
         <line-breakpoint enabled="true" type="DlvLineBreakpoint">
           <url>file://$PROJECT_DIR$/main.go</url>
-          <line>28</line>
+          <line>7</line>
           <option name="timeStamp" value="4" />
         </line-breakpoint>
       </breakpoints>
